+++ crates/template/src/test_fuzz.rs
@@ -1,10 +1,15 @@
 use quote::format_ident;
+use std::collections::HashMap;
 use syn::parse_quote;
 
 use crate::Template;
 
 impl Template {
-    pub(crate) fn test_fuzz(&mut self, lib_names: &[String]) {
+    pub(crate) fn test_fuzz(
+        &mut self,
+        program_ids: &HashMap<String, String>,
+        lib_names: &[String],
+    ) {
         let mut use_statements: Vec<syn::ItemUse> = Vec::new();
 
         let mut programs: Vec<syn::Stmt> = Vec::new();
@@ -12,7 +17,8 @@ impl Template {
         let mut input_array: Vec<syn::Ident> = Vec::new();
 
         for program in lib_names {
-            let (use_statement, program, program_variable) = process_program_entries(program);
+            let (use_statement, program, program_variable) =
+                process_program_entries(program, program_ids.get(program));
             // add to the use statements
             use_statements.push(use_statement);
             // add to the programs
@@ -69,7 +75,10 @@ impl Template {
     }
 }
 
-fn process_program_entries(lib_name: &String) -> (syn::ItemUse, syn::Stmt, syn::Ident) {
+fn process_program_entries(
+    lib_name: &String,
+    program_id: Option<&String>,
+) -> (syn::ItemUse, syn::Stmt, syn::Ident) {
     // library name as identifier
     let library = format_ident!("{}", lib_name);
     // entry name as identifier
@@ -80,10 +89,16 @@ fn process_program_entries(lib_name: &String) -> (syn::ItemUse, syn::Stmt, syn::
     // initial use statement
     let use_statement = parse_quote!(use #library::entry as #library_entry;);
 
+    // program id if present, otherwise fill with placeholder
+    let program_id = match program_id {
+        Some(address) if !address.trim().is_empty() => address,
+        _ => "fill corresponding program ID here",
+    };
+
     // program definition
     let program = parse_quote! {
         let #variable_name = ProgramEntrypoint::new(
-            pubkey!("fill corresponding program ID here"),
+            pubkey!(#program_id),
             None,
             processor!(#library_entry)
         );